Created tree structure for world items
All items in a world are stored in a tree structure, the World being the
root world item. This is implemented using the QObject parent-child
relationship. This also means that memory management is happens
automatically in the game core objects. The building block of the World
tree is the WorldItem class, which also stores metadata common to all
the world items, like objectName and displayName. WorldItem also
supplies serialization and deserialization via the toJson and fromJson
interface methods. All WorldItem subclasses must correctly reimplement
these methods. To aid in looking up sibling items in the tree by name
and type the WorldItem supplies the getWorldItem interface method.
All names in the json files must follow the camel case nameing
convention from now on.