Added the classes that represent a game or a map.
Added classes:
GameObject, Unit, Settlement, Player, MapTile, Map.
GameObject is the common ancestor of all the Game classes.
A Map can represent both a map and an ongoing game state.
Added SettlementType to world.json. A SettlementType represents the
different kinds of settlements that can appear in the world.
Other fixes and adjustments.
Note: the code does not link at the moment.