Created GameEntity, a common baseclass for storage-backed GameObjects
GameEntity is a baseclass for GameObjects which can be loaded and saved
to disk. Search paths are now managed by QDir. As a consequence for that
worlds, maps and world surfaces will have different file names and
extensions:
world: objectName.wwd
map: objectName.wmd
surface: objectName.wsd
Where objectName is the objectName of the respective GameObject.