Units can now be recruited
Created "new-style" exceptions. Exceptions will be separate classes
instead of the same class with different error codes. This makes
possible to catch only those exceptions that can be handled.
Also the log-category will be renamed to loggerName and it will be the
name of the module and the class instead of just the name of the module.