WorldSurface: abstract storage and add support for surface directories
Currently only a single kind of surface is supported, the "wsp" files.
Hovewer for development it would be useful to not have to re-package the
surface after modifying an image. To support this kind of use-case a new
`storage` type was added. To enable the support for multiple storage
types the storage-related functionality was moved into polimorphic storage
classes. The current storage code was moved into ArchiveStorage and
DirectoryStorage was added for the directory-based surfaces.
Also the WorldSurface interface was refactored, the object-based image
and texture getters were removed and instead now all images and textures
are aquired by internal-paths. These paths only make sense in the
context of the surface they belong to and should only be used in
world-specific code.