Get rid of nested field types
Fields now have a simple type. Containers are now heterogenous as
opposed to homogenous, it's no longer required for each value in a
container to have the pre-specified type. It's also no longer required
to fully specify the type of values.  These estrictions brought no real
benefit in generating bindings and imposed unecessary restrictions
on script code. Have a real dynamic type-system instead for
the component-fields.