Add io/Visitor.hpp
The infrastructure introduced by Visitor.hpp can be used to describe a
class and its members compile-time and use this knowledge to generate
serializers/unserializers. This is the first step in generating all the
current serializers/unserializers in io. The indended useage is that
serializable classes provide a static templated describe() method that
receives a visitor and calls the visitParent(), visitMember() and
visitConstructor() methods to describe itself.