ui/Render: take entity hierarchy into consideration when drawing
Don't treat all entities the same. Entities are organized into a tree
hierarchy now. When drawing the contents of a map-node, only the root
entities (those that have no parent) should be drawn directly. Children
of these root entities should be rendered as children of the root
entities, according to their position in the entity tree, such that the
resulting `QSGNode` structure reflects that of the `core::Entity` one.