ui/WorldSurface: refactor tileWidth and tileHeight, add gridSize
Warmonger is moving away from supporting tile sizes with differing width
and height. In the future only tile-sizes that have equal width and
height will be supported. To reflect this, merge these two properties of
the world surface into a single `tileSize` property. Also add the
`gridSize` property which denotes the size of subtiles a regular tile is
divided into when rendering. This property is not used for now.

This is part of the work of moving Warmonger to support rendering
hexagons by assembling individual images, that each cover only a part of
the tile. This allows for greater flexibility for the rendering code, as
well as for artists, as the map is now not rendered by assempbling
images, each covering an entire map-node, instead each map-node is
rendered by assembling an arbitrary number of sub-images.