ui/WorldSurface: parse the graphicAssets from the meta file
This contains the manifest of all the generic graphic assets the surface
contains. The format is:
"graphicAssets": {
    "assetId1": "assetPath1",
    "assetId2": "assetPath2",
    ...
}

This will be used to map aliases to concrete paths. This is part of the
effort of moving away from using direct paths to assets through the
codebase, towards an opaque asset handle.