ui/LuaWorldSurfaceRules: don't use sol::property for member variables
sol::property() is a helper designated for masquaranding a
getter()/setter() pair as a simple object property in lua code.
Passing a reference to a member variable to it confuses it, creating a
write-only property (as it somehow matches the member reference to a
setter()). When directly exposing member variables, there is no need to
use sol::property, the member variable reference can be directly passed
to sol, which will do the correct thing.