thirdparty: use the distro provided fmt package instead of bundling it
Since (on Fedora at least) the version of the fmt library is 5.2.1 the
codebase is updated to work with this version. As part of that,
utils/Format.{h,cpp} is dropped, instead of the `format_to*(` overloads
provided therein (deprecated apparently in fmt 5) we now rely on
`fmt/ostream.h` to piggyback on existing `operator<<` overloads for
printing custom types.

It is worth noting that `spdlog` bundles its own fmt copy, which is an
older version. This necessitates making sure that we don't include our
fmt before we include any spdlog header. This is not a huge problem as
we try to hide spdlog from sight anyway, relying on it as an invisible
logging backend anyway. Still it is an annoyance.