Bye warmonger exec welcome warmonger godot native lib
Remove the warmonger executable target and replace it with the godot
native (shared) library. The entry point of the game will be through the
godot executable and existing (and future) warmonger code will serve as
the game logic library, exposed to godot through the gdnative interface.

Currently the all warmonger code is compiled into this library. We will
remove, replace and refactor code incrementally in the future. For now,
`warmonger/warmonger.cpp` is transformed from an executable entry point
to a gdnative entry point, containing only the bare minimum to qualify
as such. No classes are exported yet.