warmonger: warmonger: export the Context class
To do that the Context class had to be "godotized". It now inherits from
`godot::Reference`, has an `_register_methods()` and `_init()` method.
Code that it will not need in the future, like the palettes, is removed.
Code that we plan to make use of in the future but needs a refactoring
is surrounded by `#if 0 .. #ifdef`. I know this is ugly but it will
serve well as an in-source TODO list on what needs to be translated.