mv warmonger native
The name warmonger for the native part of the warmonger codebase is both
misleading and redundant (as it is also the name of the project). The
native library does not constitute the entire project.
Rename the folder from `warmonger` to `native` to signal that this is
just the native part of the codebase as opposed to the entire project.
Also drop the native library as a dependency from the README.md. It is
not a dependency, it is an integral part of the project.